refactor(core): split graph community services
Move snapshot loading, detection, hierarchy, membership, clustering, and
selection into acyclic canonical modules behind an explicit public facade.

Shield every snapshot waiter and count waiters per loader so one cancelled
caller cannot poison joiners or remove a replacement task.
