+++ apps/api/tests/test_graph_communities_lod.py
@@ -4,10 +4,10 @@ import asyncio
 
 import pytest
 
-import sibyl_core.services.graph_communities as communities
+import sibyl_core.services.graph_community_snapshot as community_snapshot
 from sibyl_core.models.entities import Entity, EntityType, Relationship, RelationshipType
-from sibyl_core.services.graph_communities import (
-    GRAPH_RESOLUTION_OVERVIEW,
+from sibyl_core.services.graph_communities import GRAPH_RESOLUTION_OVERVIEW, GraphSnapshot
+from sibyl_core.services.graph_community_selection import (
     _build_cluster_detail_graph_from_snapshot,
     _build_overview_graph_from_snapshot,
 )
@@ -148,12 +148,12 @@ async def test_get_graph_snapshot_fetches_entities_and_edges_concurrently(
         await entities_started.wait()
         return []
 
-    communities.GRAPH_SNAPSHOT_CACHE.clear()
-    monkeypatch.setattr(communities, "_list_all_entities", fake_list_all_entities)
-    monkeypatch.setattr(communities, "_list_all_relationships", fake_list_all_relationships)
+    community_snapshot.GRAPH_SNAPSHOT_CACHE.clear()
+    monkeypatch.setattr(community_snapshot, "_list_all_entities", fake_list_all_entities)
+    monkeypatch.setattr(community_snapshot, "_list_all_relationships", fake_list_all_relationships)
 
     snapshot = await asyncio.wait_for(
-        communities._get_graph_snapshot(object(), "org-concurrent"),
+        community_snapshot._get_graph_snapshot(object(), "org-concurrent"),
         timeout=1,
     )
 
@@ -176,19 +176,19 @@ async def test_get_graph_snapshot_joins_concurrent_loads(
         await asyncio.sleep(0)
         return []
 
-    communities.GRAPH_SNAPSHOT_CACHE.clear()
-    communities.GRAPH_SNAPSHOT_LOADS.clear()
-    monkeypatch.setattr(communities, "_list_all_entities", fake_list_all_entities)
-    monkeypatch.setattr(communities, "_list_all_relationships", fake_list_all_relationships)
+    community_snapshot.GRAPH_SNAPSHOT_CACHE.clear()
+    community_snapshot.GRAPH_SNAPSHOT_LOADS.clear()
+    monkeypatch.setattr(community_snapshot, "_list_all_entities", fake_list_all_entities)
+    monkeypatch.setattr(community_snapshot, "_list_all_relationships", fake_list_all_relationships)
 
     first, second = await asyncio.gather(
-        communities._get_graph_snapshot(
+        community_snapshot._get_graph_snapshot(
             object(),
             "org-single-flight",
             max_entities=100,
             max_relationships=200,
         ),
-        communities._get_graph_snapshot(
+        community_snapshot._get_graph_snapshot(
             object(),
             "org-single-flight",
             max_entities=100,
@@ -207,7 +207,7 @@ async def test_get_graph_snapshot_cancels_inflight_load_on_request_cancel(
     started = asyncio.Event()
     cancelled = asyncio.Event()
 
-    async def fake_load_graph_snapshot(*args, **kwargs) -> communities.GraphSnapshot:
+    async def fake_load_graph_snapshot(*args, **kwargs) -> GraphSnapshot:
         started.set()
         try:
             await asyncio.sleep(10)
@@ -216,12 +216,12 @@ async def test_get_graph_snapshot_cancels_inflight_load_on_request_cancel(
             raise
         raise AssertionError("snapshot load should have been cancelled")
 
-    communities.GRAPH_SNAPSHOT_CACHE.clear()
-    communities.GRAPH_SNAPSHOT_LOADS.clear()
-    monkeypatch.setattr(communities, "_load_graph_snapshot", fake_load_graph_snapshot)
+    community_snapshot.GRAPH_SNAPSHOT_CACHE.clear()
+    community_snapshot.GRAPH_SNAPSHOT_LOADS.clear()
+    monkeypatch.setattr(community_snapshot, "_load_graph_snapshot", fake_load_graph_snapshot)
 
     task = asyncio.create_task(
-        communities._get_graph_snapshot(
+        community_snapshot._get_graph_snapshot(
             object(),
             "org-cancel",
             max_entities=100,
@@ -235,7 +235,7 @@ async def test_get_graph_snapshot_cancels_inflight_load_on_request_cancel(
         await task
 
     assert cancelled.is_set()
-    assert communities.GRAPH_SNAPSHOT_LOADS == {}
+    assert community_snapshot.GRAPH_SNAPSHOT_LOADS == {}
 
 
 @pytest.mark.asyncio
@@ -245,17 +245,17 @@ async def test_get_graph_snapshot_joiner_cancel_does_not_cancel_shared_load(
     started = asyncio.Event()
     release = asyncio.Event()
 
-    async def fake_load_graph_snapshot(*args, **kwargs) -> communities.GraphSnapshot:
+    async def fake_load_graph_snapshot(*args, **kwargs) -> GraphSnapshot:
         started.set()
         await release.wait()
-        return communities.GraphSnapshot(entities=[], relationships=[], entity_by_id={})
+        return GraphSnapshot(entities=[], relationships=[], entity_by_id={})
 
-    communities.GRAPH_SNAPSHOT_CACHE.clear()
-    communities.GRAPH_SNAPSHOT_LOADS.clear()
-    monkeypatch.setattr(communities, "_load_graph_snapshot", fake_load_graph_snapshot)
+    community_snapshot.GRAPH_SNAPSHOT_CACHE.clear()
+    community_snapshot.GRAPH_SNAPSHOT_LOADS.clear()
+    monkeypatch.setattr(community_snapshot, "_load_graph_snapshot", fake_load_graph_snapshot)
 
     owner = asyncio.create_task(
-        communities._get_graph_snapshot(
+        community_snapshot._get_graph_snapshot(
             object(),
             "org-join-cancel",
             max_entities=100,
@@ -265,7 +265,7 @@ async def test_get_graph_snapshot_joiner_cancel_does_not_cancel_shared_load(
     await started.wait()
 
     joiner = asyncio.create_task(
-        communities._get_graph_snapshot(
+        community_snapshot._get_graph_snapshot(
             object(),
             "org-join-cancel",
             max_entities=100,
